Move linear model to separate files
